test: Adjusting unit tests for @ValueMapping combinations
