+++ processor/src/test/java/org/mapstruct/ap/test/mappingcomposition/CompositionTest.java
@@ -6,15 +6,9 @@
 package org.mapstruct.ap.test.mappingcomposition;
 
 import java.util.Date;
-
-import org.mapstruct.ap.test.mappingcomposition.value.CustomValueAnnotation;
-import org.mapstruct.ap.test.mappingcomposition.value.ValueMappingCompositionMapper;
-import org.mapstruct.ap.test.value.ExternalOrderType;
-import org.mapstruct.ap.test.value.OrderType;
 import org.mapstruct.ap.testutil.IssueKey;
 import org.mapstruct.ap.testutil.ProcessorTest;
 import org.mapstruct.ap.testutil.WithClasses;
-import org.mapstruct.factory.Mappers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -71,35 +65,4 @@ public class CompositionTest {
 
     }
 
-    @ProcessorTest
-    @WithClasses({
-        ValueMappingCompositionMapper.class,
-        ExternalOrderType.class,
-        OrderType.class,
-        CustomValueAnnotation.class
-    })
-    public void shouldValueCompositionSuccess() {
-        ValueMappingCompositionMapper compositionMapper = Mappers.getMapper( ValueMappingCompositionMapper.class );
-        assertThat( compositionMapper.orderTypeToExternalOrderType( OrderType.EXTRA ) )
-            .isEqualTo( ExternalOrderType.SPECIAL );
-        assertThat( compositionMapper.orderTypeToExternalOrderType( OrderType.NORMAL ) )
-            .isEqualTo( ExternalOrderType.DEFAULT );
-    }
-
-    @ProcessorTest
-    @WithClasses({
-        ValueMappingCompositionMapper.class,
-        ExternalOrderType.class,
-        OrderType.class,
-        CustomValueAnnotation.class
-    })    public void duplicateValueMappingAnnotation() {
-        ValueMappingCompositionMapper compositionMapper = Mappers.getMapper( ValueMappingCompositionMapper.class );
-        assertThat( compositionMapper.duplicateAnnotation( OrderType.EXTRA ) )
-            .isEqualTo( ExternalOrderType.SPECIAL );
-        assertThat( compositionMapper.duplicateAnnotation( OrderType.STANDARD ) )
-            .isEqualTo( ExternalOrderType.SPECIAL );
-        assertThat( compositionMapper.duplicateAnnotation( OrderType.NORMAL ) )
-            .isEqualTo( ExternalOrderType.DEFAULT );
-
-    }
 }
