+++ processor/src/test/java/org/mapstruct/ap/test/value/composition/ValueMappingCompositionMapper.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.mappingcomposition.value;
+package org.mapstruct.ap.test.value.composition;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ValueMapping;
